fixed includes (using iwyu)
I've run the 'iwyu' tool on the code, and this is the result.

The code seems in fact to build with the suggested modifications, although
I only checked it under GNU.  The plan is to set up a little build
environment with the other supported OS, as we discussed by mail.

Unrelated: I'm wondering what system requires sys/malloc.h to be included
instead of stdlib.h.

Signed-off-by: Johannes von Rotz <jr@vrtz.ch>