Avoid exposed global
The use of globals is not recommendable in general, but it's a bit better
if they're at least only visible from within the module they're used by.

Later on we may get rid of the globals entirely.  For the moment it is
enough to avoid copying around an option into a global buffer.

Signed-off-by: Johannes von Rotz <jr@vrtz.ch>